requirements1.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
crispy==0.7.4
crispy-bootstrap5==0.7
cryptography==41.0.2
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.20
django-allauth==0.54.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-phonenumber-field==7.1.0
extension==1.0.0
fabio==2023.6.0
Flask==2.2.5
fonttools==4.38.0
frozenlist==1.3.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.3
multidict==6.0.4
numpy==1.21.6
oauthlib==3.2.2
packaging==23.1
phonenumberslite==8.13.16
Pillow==9.5.0
plotly==5.15.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
silx==1.1.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
twilio==8.5.0
typing_extensions==4.7.1
urllib3==1.26.6
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.15.0
